Split josh-gui into per-page source files
Change: split-gui-modules
